vis/requirements.txt
flask
psycopg2-binary
plotly
